fix: Add numeric type casting to ROUND function calls in mv_revenue_forecast
- Cast m.mrr_usd to numeric in ROUND calls
- Cast multiplication and division results to numeric
- Ensures PostgreSQL ROUND function signature compatibility

+++ packages/db/migrations/0022_reportes_remaining_mvs.sql
@@ -159,10 +159,10 @@ pipeline_value AS (
 )
 SELECT
   m.tenant_id,
-  ROUND(m.mrr_usd, 2)                               AS mrr_usd,
-  ROUND(m.mrr_usd * 12, 2)                          AS arr_usd,
+  ROUND(m.mrr_usd::numeric, 2)                      AS mrr_usd,
+  ROUND((m.mrr_usd * 12)::numeric, 2)               AS arr_usd,
   ROUND(
-    m.mrr_usd / NULLIF(c.monthly_churn_rate, 0), 2
+    (m.mrr_usd / NULLIF(c.monthly_churn_rate, 0))::numeric, 2
   )                                                  AS ltv_usd,
   ROUND(c.monthly_churn_rate::numeric, 4)            AS monthly_churn_rate,
   COALESCE(ROUND(pv.weighted_pipeline_usd::numeric, 2), 0) AS weighted_pipeline_usd,
